+++ packages/page-accounts/src/modals/MultisigCreate.tsx
@@ -1,15 +1,15 @@
-// Copyright 2017-2021 @polkadot/app-accounts authors & contributors
+// Copyright 2017-2022 @polkadot/app-accounts authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
 import type { ActionStatus } from '@polkadot/react-components/Status/types';
 import type { ModalProps } from '../types';
 
-import BN from 'bn.js';
 import React, { useCallback, useState } from 'react';
 
 import { Button, Input, InputAddressMulti, InputNumber, Modal } from '@polkadot/react-components';
 import { useApi } from '@polkadot/react-hooks';
 import { keyring } from '@polkadot/ui-keyring';
+import { BN } from '@polkadot/util';
 
 import useKnownAddresses from '../Accounts/useKnownAddresses';
 import { useTranslation } from '../translate';