+++ packages/page-signing/src/index.tsx
@@ -1,4 +1,4 @@
-// Copyright 2017-2021 @polkadot/app-signing authors & contributors
+// Copyright 2017-2022 @polkadot/app-signing authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
 import type { AppProps as Props } from '@polkadot/react-components/types';